requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0

# Detection & Vision
ultralytics>=8.1.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# CLIP & Transformers
clip-by-openai>=1.0
transformers>=4.35.0
huggingface-hub>=0.19.0

# Data & Analysis
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
click>=8.1.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
